refactor(structure): 📦 Reorganize project structure by moving app subdirectories to src
- Move src/app/providers → src/providers
- Move src/app/layouts → src/layouts
- Move src/app/store → src/store
- Move src/app/router → src/router
- Update all import paths from @/app/* to @/*
- Remove empty src/app directory

This creates a cleaner, flatter structure with all major concerns
directly accessible from the src directory.
