+++ src/App.tsx
@@ -1,5 +1,5 @@
 import { RouterProvider } from 'react-router-dom'
-import { router } from '@/app/router'
+import { router } from '@/router'
 import './App.css'
 
 function App() {
