+++ src/hooks/useThemeColor.ts
@@ -1,5 +1,5 @@
 import { useEffect, useMemo, useState } from 'react'
-import { useTheme } from '@/app/providers/useTheme'
+import { useTheme } from '@/providers/useTheme'
 
 // Global MutationObserver instance to prevent multiple observers
 class GlobalThemeObserver {
