+++ src/layouts/Header.tsx
@@ -10,7 +10,6 @@ import {
   User,
 } from 'lucide-react'
 import { useState } from 'react'
-import { useTheme } from '@/app/providers/useTheme'
 import { Avatar, AvatarFallback, AvatarImage } from '@/components/ui/Avatar'
 import { Badge } from '@/components/ui/Badge'
 import { Button } from '@/components/ui/Button'
@@ -26,6 +25,7 @@ import {
   DropdownMenuTrigger,
 } from '@/components/ui/DropdownMenu'
 import { Input } from '@/components/ui/Input'
+import { useTheme } from '@/providers/useTheme'
 import { ComponentErrorBoundary } from '@/shared/components/ErrorBoundary'
 
 interface HeaderProps {
